requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
django-celery-results
python-dotenv
gunicorn